// src/core/models/AudioData.ts
export interface AudioData {
    audio_files: AudioFile[];
}

export interface AudioFile {
    id: number;
    chapter_id: number;
    file_size: number;
    format: string;
    audio_url: string;
    duration: number;
    verse_timings: VerseTiming[];
}

export interface VerseTiming {
    verse_key: string;
    timestamp_from: number;
    timestamp_to: number;
    duration: number;
    segments: number[][];
}